refactor(ingestor): group run-* command orchestrators into commands/ subfolder
Domain slices (ebird/, inat/, r2/, archive/, wikipedia/) were already
foldered. The run-* tier was the one un-grouped seam — 8 job orchestrators
living flat at src/ root alongside cli.ts, handler.ts, and transform.ts.
Moving them into src/commands/ gives each layer a clear home and makes
navigation ("where are the schedulable jobs?") obvious.

Zero logic changes. Import paths in callers (cli.ts, handler.ts, index.ts,
cli.test.ts, handler.test.ts) updated to ./commands/run-*.js; imports inside
the moved files updated one level up (../ebird/, ../transform.js, etc.).
Monitoring.tf doc-comment updated to reflect the new path.
