+++ services/ingestor/src/commands/run-hotspots.test.ts
