+++ services/ingestor/src/commands/run-photos.test.ts
