+++ services/ingestor/src/index.ts
@@ -1,6 +1,6 @@
 export { handleScheduled, type HandlerEnv, type ScheduledKind } from './handler.js';
-export { runIngest, type RunSummary } from './run-ingest.js';
-export { runHotspotIngest, type RunHotspotSummary } from './run-hotspots.js';
-export { runBackfill, type RunBackfillSummary } from './run-backfill.js';
-export { runPhotos, type RunPhotosSummary, type RunPhotosArgs } from './run-photos.js';
+export { runIngest, type RunSummary } from './commands/run-ingest.js';
+export { runHotspotIngest, type RunHotspotSummary } from './commands/run-hotspots.js';
+export { runBackfill, type RunBackfillSummary } from './commands/run-backfill.js';
+export { runPhotos, type RunPhotosSummary, type RunPhotosArgs } from './commands/run-photos.js';
 export { fetchInatCandidates, type InatCandidate, type DenyContext } from './inat/candidates.js';
